kernel: Provide cfg for users to enable kernel CONFIG options
Yocto provides various mechanisms for manipulating kernel config
including config fragments in .cfg files and the kernel_configure_variable
syntax in an include file.

For users not so familiar with Yocto or those who prefer the cfg
mechanism and simply have a few kernel CONFIG options they need to
enable provide a user.cfg file they can put them in. Then add it
into the SRC_URI so they will be included in the build.